requirements.txt
Django==6.0
django-environ==0.12.0
psycopg[binary]==3.3.2
